Fixed Vector of type Double to Float

+++ csc133-object-oriented-computer-graphics-programming/lab/lab01-SkyMail3000-part01/src/org/csc133/a1/GameObject.java
@@ -32,7 +32,7 @@ public abstract class GameObject {
      */
 
     // All game objects have a location
-    private Vector<Double> position;
+    private Vector<Float> position;
     // non-negative values X initially should be in the range 0.0 to 1024.0
     public static final int WORLD_WIDTH = 1024;
     // non-negative values Y initially should be in the range 0.0 to 768.0
